added auto find lower bound for rat ii modulo scheduler

+++ src/HatScheT/base/ModuloSchedulerBase.h
@@ -41,7 +41,7 @@ public:
    * \brief getMinII
    * \return
    */
-  int getMinII(){return this->minII;}
+  double getMinII(){return this->minII;}
   /*!
    * \brief getMaxII
    * \return
